actually use libvr4300
